moved epoch tracker queries
