+++ x/stakeibc/keeper/grpc_query.go
@@ -181,6 +181,34 @@ func (k Keeper) AddressUnbondings(c context.Context, req *types.QueryAddressUnbo
 	return &types.QueryAddressUnbondingsResponse{AddressUnbondings: addressUnbondings}, nil
 }
 
+func (k Keeper) EpochTrackerAll(c context.Context, req *types.QueryAllEpochTrackerRequest) (*types.QueryAllEpochTrackerResponse, error) {
+	if req == nil {
+		return nil, status.Error(codes.InvalidArgument, "invalid request")
+	}
+
+	ctx := sdk.UnwrapSDKContext(c)
+	epochTrackers := k.GetAllEpochTracker(ctx)
+
+	return &types.QueryAllEpochTrackerResponse{EpochTracker: epochTrackers}, nil
+}
+
+func (k Keeper) EpochTracker(c context.Context, req *types.QueryGetEpochTrackerRequest) (*types.QueryGetEpochTrackerResponse, error) {
+	if req == nil {
+		return nil, status.Error(codes.InvalidArgument, "invalid request")
+	}
+	ctx := sdk.UnwrapSDKContext(c)
+
+	val, found := k.GetEpochTracker(
+		ctx,
+		req.EpochIdentifier,
+	)
+	if !found {
+		return nil, status.Error(codes.NotFound, "not found")
+	}
+
+	return &types.QueryGetEpochTrackerResponse{EpochTracker: val}, nil
+}
+
 func (k Keeper) AllTradeRoutes(c context.Context, req *types.QueryAllTradeRoutes) (*types.QueryAllTradeRoutesResponse, error) {
 	if req == nil {
 		return nil, status.Error(codes.InvalidArgument, "invalid request")
